Refactor Persona and Avatar Components into ChooseCharacter Component
