+++ src/components/ui/ChooseCharacterDisplay/AvatarGallery/AvatarGallery.cy.tsx
