+++ src/components/ui/ChooseCharacterDisplay/AvatarGallery/AvatarGallery.tsx
