+++ src/components/ui/index.ts
@@ -1,6 +1,6 @@
 // Barrel File
 export * from './AuthModal/AuthModal';
-export * from './AvatarGallery/AvatarGallery';
+export * from './ChooseCharacterDisplay/AvatarGallery/AvatarGallery';
 export * from './HexagonCollection/HexagonCollection';
 export * from './HexagonItem/HexagonItem';
 export * from './InfoModal/InfoModal';
@@ -10,7 +10,7 @@ export * from './Layout/TwoColumnLayout';
 export * from './LinkCard/LinkCard';
 export * from './Loading/Loading';
 export * from './Navigation/Navigation';
-export * from './PersonaList/PersonaList';
+export * from './ChooseCharacterDisplay/PersonaList/PersonaList';
 export * from './ProgressTracker/ProgressTracker';
 export * from './RichTextOutput/RichTextOutput';
 export * from './Settings/Settings';
